Added code to close client session with server after GUI closed.
Removed INFORM_OF_STALEMATE and INFORM_SHUT_DOWN from ServerToClientMessageType enum class.

Removed Client and Game attributes from ClientMessageToServer as not required.

Added Session object to ClientMessageToServer.

Made it so that the Session id is no longer passed to the server in the establishConnection() method, as this can be obtained from the Session object.

Updated README.